requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Apache Spark
pyspark==3.5.0

# Configuration Management
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database Connectivity
pymssql==2.2.11
mysql-connector-python==8.2.0

# AWS SDK
boto3==1.34.0

# Utilities
python-dateutil==2.8.2
